Рефакторинг работы с пользователем

+++ src/BulletinBoard/Application/BulletinBoard.Application.AppServices/Contexts/User/Services/UserService.cs
@@ -2,13 +2,13 @@
 using BulletinBoard.Contracts.User;
 using BulletinBoard.Contracts.Users;
 using System.Linq.Expressions;
+using System.Threading;
 
 namespace BulletinBoard.Application.AppServices.Contexts.User.Services
 {
     /// <inheritdoc/>
     public class UserService : IUserService
     {
-        // private readonly IUserRepository _userRepository;
         private readonly IUserRepository _userRepository;
 
         /// <summary>
@@ -20,9 +20,9 @@ namespace BulletinBoard.Application.AppServices.Contexts.User.Services
             _userRepository = userRepository;
         }
 
-        public Task<IEnumerable<UserDto>> GetAllAsync()
+        public Task<IEnumerable<UserDto>?> GetAllAsync(CancellationToken cancellationToken)
         {
-            return _userRepository.GetAllAsync();
+            return _userRepository.GetAllAsync(cancellationToken);
         }
 
         public Task<UserDto?> GetByIdAsync(Guid id, CancellationToken cancellationToken)
@@ -30,34 +30,41 @@ namespace BulletinBoard.Application.AppServices.Contexts.User.Services
             return _userRepository.GetByIdAsync(id, cancellationToken);
         }
 
-        public Task<UserDto> GetFirstWhere(Expression<Func<Domain.User, bool>> predicate, CancellationToken cancellationToken)
+        public Task<UserDto?> GetFirstWhere(Expression<Func<Domain.User, bool>> predicate, CancellationToken cancellationToken)
         {
             return _userRepository.GetFirstWhere(predicate, cancellationToken);
         }
 
-        public Task<IEnumerable<UserDto>> GetRangeByIDAsync(List<Guid> ids, CancellationToken cancellationToken)
+        public async Task<Guid> AddAsync(CreateUserDto user, CancellationToken cancellationToken)
         {
-            return _userRepository.GetRangeByIDAsync(ids, cancellationToken);
+            var existedUser = await _userRepository.GetFirstWhere(u => u.Email == user.Email, cancellationToken);
+            if (existedUser != null)
+            {
+                return Guid.Empty;
+            }
+
+            return await _userRepository.AddAsync(user, cancellationToken);
         }
 
-        public Task<IEnumerable<UserDto>> GetWhere(Expression<Func<Domain.User, bool>> predicate)
+        public async Task<bool> UpdateAsync(Guid id, EditUserDto user, CancellationToken cancellationToken)
         {
-            return _userRepository.GetWhere(predicate);
+            var existedUser = await _userRepository.GetByIdAsync(id, cancellationToken);
+            if (existedUser == null)
+            {
+                return true;
+            }
+
+            return await _userRepository.UpdateAsync(id, user, cancellationToken);
         }
 
-        public Task<Guid> AddAsync(CreateUserDto user, CancellationToken cancellationToken)
+        public async Task<bool> DeleteAsync(Guid id, CancellationToken cancellationToken)
         {
-            return _userRepository.AddAsync(user, cancellationToken);
-        }
-
-        public Task<bool> UpdateAsync(Guid id, EditUserDto user, CancellationToken cancellationToken)
-        {
-            return _userRepository.UpdateAsync(id, user, cancellationToken);
-        }
-
-        public Task<bool> DeleteAsync(Guid id, CancellationToken cancellationToken)
-        {
-            return _userRepository.DeleteAsync(id, cancellationToken);
+            var existedUser = await _userRepository.GetByIdAsync(id, cancellationToken);
+            if (existedUser == null)
+            {
+                return true;
+            }
+            return await _userRepository.DeleteAsync(id, cancellationToken);
         }
     }
 }
